depth/Unidepth/requirements.txt
appdirs
attrs
black
blosc2
botocore==1.34.54
certifi==2022.12.7
charset-normalizer
click
contourpy
cycler
docker-pycreds
einops==0.7.0
filelock
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0
fonttools
fsspec
fvcore==0.1.5.post20221221
gitdb
GitPython
h5py>=3.10.0
huggingface-hub>=0.22.0
idna
imageio
imath
iopath
isort
Jinja2
jmespath
kiwisolver
MarkupSafe
matplotlib
mccabe
mpmath
msgpack
mypy-extensions
ndindex
networkx
ninja
numexpr
numpy<2.0.0
opencv-python
OpenEXR
packaging
pandas
pathspec
pillow==10.2.0
platformdirs
portalocker
protobuf==4.25.3
psutil
py-cpuinfo
pycodestyle
pyflakes
pyparsing
python-dateutil
pytz
PyYAML
requests
safetensors
scipy
sentry-sdk
setproctitle
six
smmap
sympy
tables
tabulate
termcolor
timm
tqdm
#triton==2.2.0
typing_extensions
tzdata==2024.1
urllib3==1.26.13
wandb
yacs
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
xformers==0.0.24